// src/DockiUp.Frontend/src/app/api/model/projectFilesDto.ts
/**
 * DockiUp.API | v1
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ProjectFileEntryDto } from './projectFileEntryDto';


export interface ProjectFilesDto { 
    path: string;
    isGit: boolean;
    composeFile: string | null;
    entries: Array<ProjectFileEntryDto>;
}